feat: reserve charm revision
Adds the reserve charm revision to the service. It is expected that
we'll want to add additional arguments to the method, but I'm waiting
to fully implement the work flow before adding everything. I suspect
we can just add a small essential metadata. In that case, we might
want to have just a EssentialCharm type that represents just that.