chore: commit message contributing guidelines
Now that we use Conventional Commits[1] for commit messages, add
that to the contributing guidelines.

In addition, clean up some of the wording that is required (local
cloning can be found elsewhere) and duplication of existing content.

Ensure that we're correct when updating the dependencies (using
tags and fallback commit hashes).

 1. https://www.conventionalcommits.org/en/v1.0.0/