chore: api/agent context aware
Threads through the context.Context for every method in the
api/agent package. This is a long an laborious set of manual changes.
The motivation behind this, is to allow observability from all agents
to the controller. With the premise being to better understand how
calls interact with the server. This can help us better diagnose
the behaviour in the system and hopefully identify and fix bugs or
performance issues.

The client and controller api packages will follow in due course.

All of the code changes are mechanical, they're just either passing
in a context.Context from the parent caller or creating a scoped
context for things like workers.