feat: Late determine container networking.
This commit attempts to introduce a new way for calculating the
value for container networking method. Pre this change there existed
state logic that would intercept the model config value on the way into
the database and attempt to modify what the user may have set to what
this value could be later for the facade.

For example the auto determine value was "", with other possible values
being local and provider. If the value set was "" the state layer would
stand up an environ and ask if the environ supported container
addresses. In turn this would modify the value in model config to
reflect provider support.

The problems we have with this approach in that it was only ever done on
initial set being either at bootstrap time or model creation. It was
never retroactively handled with change to model config. It also
modifies the values from the what the user may or may not have set
taking away from the user interface aspect of model config.

Essentially we are trying to set in model config what the exact value
should be when returned to an agent through the facade instead of just
taking the value and making a determination at the facade value. i.e the
facade and model config value are highly coupled.

As we move to DQlite we have been looking at this code trying to figure
out how to move it over. We initially tried driving the value via model
defaults and the provider. The problem with this is we end up in a
chicken and egg problem where the provider needs model config to stand up
and we need the value from the provider to set model config.

The better approach that this commit aims to start moving towards is to
have the facade and service layers interrupt the value on read and
translate this into what the facade caller needs.

With this approach we de-couple model config from the facade for the
agent. We don't need to do expensive provider stuff for every delta on
model config. Lastly we get rid of special chicken and egg problems that
have been danced around in the state layer for years.