fix: failing git version check
TestMaybeGenerateVersionStringUsesAbsolutePathGitVersion fails if
the git binary isn't correctly patched.

The code for checking each version wasn't ideal. The code was
dependant on the code inside the function and used the export
tests to patch a git check. Instead we now can pass in a version
reader to check each vcs. This should improve testability of the
code path.