fix: ensure we release state
The controller dashboard leaks the statepool. Essentially it's missing
the call to release. I wouldn't expect it to be called that often, so
probably not the source of our leaks.

In addition the call to the SystemState can fail and also doesn't call
system state release. We should release it on error.

Finially, I've removed the defer and explicitly called release when
we're in a error state.