fix(cmr): expose all controller addresses cross-controller
We were previously sourcing addresses for cross-controller info from
APIHostPortsForAgents and prioritising for local-cloud scope.

This meant that we were not exposing public addresses resulting in some
CMRs not being able to contact remote controllers.

Here we user APIHostPortsForClients, which includes public addresses,
and we prioritise for the public scope.