fix: allow migration from 3.6
The restriction within the model service prevented the migration
from 3.6 edge. We don't want that restriction to exist whilst 3.6
is an ongoing LTS. The LTS must be able to migrate to main and we
don't want to be in a mad scramble to fix it. This change will
mandate that a migration from 3.6 to main must work from the get go.

No exceptions.