feat(authorisedkeys): Moves keyupdater to DQlite.
This commit moves the agent KeyUpdater facade over to the new DQlite
logic. With this change all of the business logic that once lived in
this facade has now been shifted a layer down into the service this
facade relies upon.

The facade is now able to just ask for the authorised keys for a machine
id and not have to worry about where this information lives or comes
from.

I have also used this facade to demonstrate a new form of error handling
for the facade. Errors encounted by this facade have been divided into
two categories. First is errors called rpc errors that are internal
server errors that occur unexpectedly. These are immediately returned
and fail the whole call.

The second type of error is param errors where the error is one
understood and is a problem with the data provided by the user. These
errors are well constructed in the facade and not deferred to a generic
catch all error handler. These strong error returns are documented and
tested for.

In this commit I have also changed the testing of the facade to use
mocks and no longer require apiserver suite.