refactor: Refactor of model id.
In previous commit I introduced the concept of model id to represent the
unique id given to machines in Juju. We don't currently use this id
outside of the domain layer.

We do however require a way to represent the strongly typed name for a
machine that we use to work on machines. This commit changes ID to Name
as a strong type.