feat: introduces service layer for key updater
This commit introduces a new service layer called keyupdater. The key
updater is responsible for helping update machines in Juju with new
authorised keys that should be propogated on to a Juju machine within a
model.

The service provides methods for listing the current keys that should be
present on a machine and for also watching authorised key updates for a
machine.

Previously this information was contained within model config and it was
the keyupdater facades responsability to pull apart model config and
amalgamate this information with the state public keys of the
controller. All of this logic has now been moved down into the service
allowing the facade to just ask for the keys for a machine.

Tests have been constructed to make sure that we are retaining the
expected buisness logic originally found in the facade.