fix: migrate 3.6 to 4.0
We must ensure that 3.6 can migrate to 4.0 without any issues once
it lands into the 4.0 controller. 3.6 is expected to be our LTS and
4.0 is going to be our new shiny. We must always ensure that we
can migrate from there to 4.0.

Solving this is done to restore the errors from the APICall to ensure
it's converted back to NotImplemented error (along with other errors).
We don't do it for all results, but just the calls to the facade, as
that's when it caught. For consistency I've added to all uniter calls,
with the aim that in new versions of Juju the API client calls should
always do this.

The SLALevel and MeterStatus are both removed in 4.0[1] and so we
need to handle when that occurs. The best way is to handle the error
and then allow call sites decide what to do.

Lastly, the meter-status and metric-sender workers will just complain
and complain about the issue. So to prevent that, we trap the error
in the manifold filter function and return nil. This will mark the
worker as complete and won't come up again unless requested. We can't
use dependency.ErrUninstall as other transitive dependencies start to
bounce and chaos ensues.

The real solution to the problem is to then perform upgrade-model,
which will then move it to 4.0.x and everything will become normalized.

 1. https://github.com/juju/juju/pull/16942