feat: charm domain service - set charm
This follows on from the initial scaffolding PR[1], by adding all
the encoding logic to set a charm. This ensures that we can encode
and decode the charm metadata, manifest actions, config and lxd
profile data.

By making parts of the charm data optional when setting it, the
plan is to drop setting essential metadata. We will have all the data
we need to make a successful requestion.

Landing this then unlocks the state implementation. Wire up of state
can then come later.

 1. https://github.com/juju/juju/pull/17569