refactor: get space related subnet fields directly from view
When getting a space and all its subnets, there are some subnet fields
that are space information, such as the space name or uuid. We were
setting these directly in the go code from the space information, but
since the information is in the view, we might as well just get it from
the database since it is being retrived anyway.