fix: errors and use sqlair in upgrade domain
This commit changes the upgrade domain to use SQLair, but some other
issues were also found with it while changing.

For one, database errors were escaping from state and being caught at
the service layer. It makes a lot more sense to catch these errors next
to the queries that caused them, especially when the function has
multiple queries in it. I have moved all of the catching of these errors
onto the actual state.

There were also some comments that didn't match the error that was given
out. I have reworked the errors to match what I believe to be the
intended behaviour and fixed the callers of the function.

The generic NotFound error was also used here. I have changed that to a
specific upgradeerorros.NotFound.

There were also some tests for SetControllerDone that appeared incomplete.
I have fixed these and made behaviour expectations consistant across the
comments.