feat: send model description on migration
To be improve the robustness of model migration, we need to better
understand what is being offered before attempting to migrate. By
sending the model description to the target controller, we can
fully understand what is on offer, we can run better pre-checks and
can give better and detailed responses based on that.

This patch will start to send the information for Precheck and
ModelInfo. It is then expected in 4.0 to be able to make use of that
information.

This is currently optional, at a later point we can enforce that
information is always present.