feat: model description pre-check
The following is a prototype that will need to be finished for
accepting the model description in the pre-check. Tests will need
to be finished, but the essential parts are functional.

It should be now feasible to work on the assumption that the model
description will be passed into the pre-check. If the description is
present then it should validate the parts it requires to mandatory.