feat(dqlite): temporarily wire up modelmachinewatcher in instancepoller api on apiserver
This is because we currently don't have the necessary functions in the
machine domain to have the instancepoller worker to use the machine
service directly instead of going through the api. Once we have those,
then we can rip out the modelmachinewatcher from the instancepoller
api like we did in the firewaller.