fix: data race in rpc layer
When attempting to read the tombstones to see if a request was
cancelled, there is a potential to a data race. Not only is there
a potential data race, I was able to witness it. The real issue is
that I was only able to observe the race in the machine logs, not
via the debug-logs. Concerning, as what else are we missing out on?