feat: use controller uuid for controller access permissions
The controller access permission used the controller namespace
"controller" as the target of the permission. For all other permissions
the UUID of the target entity is used. It should be the same here.

The controller access permission is changed to use the controller UUID.
This has required piping the UUID through all the appropriate places.

Fix SQL