feat: charm domain state
Adds the beginnings of the state methods for the charm state. This
is tackling all the simple getters and setter methods. The more
complex getters will follow.

Unfortunately, I ran into a sqlair bug, so had to change tack and
ensure that we could get the right information out that we wanted.
This meant that we stuck to the error conditions. If the charm
doesn't exist we return NotFound, rather than the default case.

 1. https://github.com/canonical/sqlair/issues/154