fix(machine): only destroy controller machines with force
We have an uncoherent behavior when destroying machines:
In a model machine with a deployed unit, a message is displayed stating that it
needs to be destroyed with --force because a unit is deployed. On a
controller machine (with the controller unit deployed on it) we also get the
displayed message but then it gets destroyed without passing --force.

The fix here is simply to remove a path that ensured graceful cleanup of
the controller unit in the case it was a controller machine. Thus, the
resulting path will be the same both for model and controller machines.