feat: add get charm metadata
Getting the charm metadata from the database is quite involved. We
have to select from a lot of tables to populate the metadata. This
is really wasteful if the code only wants the relations. It is
also not time efficent to get all the metadata as row based type. If
a charm metadata is fully populated it will result in thousands of
rows to populate. It was then decided to not use row based requests
for getting metadata. Instead we'll fallback to multiple selects.
The advantage of doing it this way, is that we can reuse the
individual methods to expose calls to get just releations or
storage.

There is a potential problem with arrays. Duplicates of values are
possible in the charm metadata and we don't prevent that in any
part of the library. Also tables are sets of rows, not ordered
rows based on insertion time. We should come up with a strategy
to keeping order of arrays. Currently, the best thought is to
just insert each row with an index, perserving order correctly.