feat: improve error messages in switch command
Before, the error messages in the switch.go file lacked specifics regarding the source of errors
commonly related to flags and arguments. This led to less clear error feedback for users.

After this commit, error messages that are returned when invalid flags or arguments are used have
been made more explicit and meaningful, improving debugging for both developers and users.

Signed-off-by: Guillaume Fouillet <guillaume.fouillet@canonical.com>