feat(domain): add machine_requires_reboot table in dqlite
Before this commit, the machine_requires_reboot table was not present in dqlite schema.
After this commit, a new table machine_requires_reboot has been added to the dqlite schema, thus
extending the capability to track and manage machines that require a reboot operation. This change
is also reflected in the schema test.go file where the new table is included in the list.

It is the first step of migration between mongodb and dqlite for
handling Machine reboot.

Signed-off-by: Guillaume Fouillet <guillaume.fouillet@canonical.com>