fix(tests): fix deploy pre deployed test
We recently removed series from the charm url when we deploy local
charms.

This means pre-deployed charms are to be re-deployed with different
urls.

In one of our tests, we hard-coded the charm url to deploy a pre-deplyed
charm. This is now incorrect, and caused the test to fail

Fix this by removing the series from the charm url