feat: larger back-off strategy
When looking over the parca results, we spend a _lot_ of time, just
polling the database for changes. For complex models with lots of
changes having 100 millisecond back-off allows for changes to be
coalesced into one term. Yet for models with zero or no changes, then
we should allow the change stream to back-off even more. If there
are no changes, we shouldn't attempt to poll.

The theory stands, that models with no activity will get short
concentrated bursts of activity, before backing off. Coalescing
is not such an need, but is desirable, as opposed to active models
with a substained throughput of changes will need to take advantage
of coalescng. The former will more likely have resources to deal with
short bursts of activity, where as coalescing on active models will
allow them to deal with the pressure.

The code currently resets the attempts to 0 upon any changes, allowing
to start polling more frequently. It maybe more beneficial to reduce
the attempt by half instead of 0. Thus, we don't hit the database
for all models at the same time in case there is an outside force
coordinating the change (clock synchronisation point for example).
This would give us more of a sinusoidal wave rather than the more
abupt sawtooth wave. The former is more compassionate and forgiving
on the surrounding infrastructure. The downside is that it _may_
take longer to get changes (we're talking 100s milliseconds in
reality).