fix: assert user auth tag
The returning auth tag from the api open call should be validated
against the one the user supplied. This can especially happen when
using the external SSO and the user suplies a different user tag
from the one they login with.

The fix is to validate that information and ensure it matches. If
it doesn't match prevent the login sending back an unauthorized
error. This will still keep the existing login details in the
accounts.yaml.