feat: get manifest
Get the manifest bases from the state and decodes them correctly.
The manifest bases should be correctly ordered because of the stored
index, although we do run into another sqlair bug[1].

Architectures will not be guaranteed to be in order, because we're
storing the base array and the architectures as one flat row. If we
did want to do that, we should create another table that kept track
of that one as well. For now, order of architectures shouldn't matter
as we don't pick one over the other.

 1. https://github.com/canonical/sqlair/issues/155