feat: remove upgrade series/base
This changeset removes all of the upgrade series/base work. We're
not going to transition this work over to 4.0, so we can just remove
it. This may be put back in the future, but it won't be done for
the initial release of 4.0.

The additional benefit to the removal of the work is that it does
simpify the machine and charm epics.

Work to ensure that 3.6 agent doesn't crash when faced with missing
charms will also need to be done. That work will be done against the
3.6 branch.