feat(migration): prevent pre-4.0 models with fan config to be migrated
This patch adds the pre-checks on the target controller (+4.0) for
validating that no fan networking is imported into it.

These are the two checks on model configs:
- Make sure fan-config is empty.
- Make sure container-networking-method is not "fan".

The second check was added only as a safeguard, because normally you
cannot set it without setting fan-config to some value. And the
fan-config being empty is the first check we perform when migrating.

This was done thanks to the work introduced in
https://github.com/juju/juju/pull/17623 which provides the model
description on migration prechecks.