feat: set metadata via set charm
This implements the metadata persistence for set charm. This includes
encoding to all the intermediate lookup tables. The LXD profile is
also writen as part of the metadata.

We can't use the column name "index" as sqlair bug prevents us from
encoding that correctly and instead "array_index" for the column name
states what it is.

Tests to ensure that we can encode and decode the same metadata have
been added, although total coverage with all edge cases might be
hard to acheive.

Work for deletion of a charm and all of it's other tables will come
later.