refactor: add param types for ssh list
Listing a users ssh keys in Juju allows the caller to specify if they
want the full key in the api list or the short fingerprint. We used a
string enum to represent this flag via params. The string enum was
defined in the juju/utils/ssh package and being used on the API params.

The problem with this approach is we have an external package that isn't
versioned with our API defining enums that are not in sync and out of
scope.

This commit brings the enums back into the params package where they
belong and removes one less need for juju/utils.