feat: key manager facade to dqlite
This commit wires up the KeyManager facade to using DQlite backed
services. This has allowed us to remove the business logic from the
facade and maintain a simple translation layer.

In this commit the keymanager facade is now making a better effort in
returning param based errors to the user.

Have also documented a current bug in the implementation where we can
not rely on the user being set in the params and have to use the
authenticated Juju user.