chore(watchers): add generic test watcher including gocheck
This commit creates a new generic struct including a watcher and a
gocheck struct, used for testing watchers. The goal here is to remove
different very similar watcher structs for testing (strings, secrets,
etc), and avoid having new ones in the future.

As part of the replacement, secret and secretbackend watcher tests have
been updated using the newly created (generic) WatcherC, and removed the
similar test watchers. Replacement of the strings watcher will be done
in a future commit.