ci: fix migrate action on main
The "Migrate" GitHub action was failing intermittently due to a timing
issue. Essentially, it takes some time after the controllers/models/
machines are deployed for them to be ready to migrate. There was
previously a hardcoded wait of 20 seconds, but this is not always
enough.

The fix is to wait for everything deployed to reach a steady state
before we begin the migration. As `juju wait-for` appears to have been
removed in 4.0, I added a Bash script `.github/waitfor.sh` that allows
us to wait on `juju status` to reach a certain state (defined by a jq
query). We have to wait on:
- the source machine to become "started"
- the ubuntu/0 unit to become "idle"
- the target controller machine to become "running"

Drive-bys:
- Now that the "Generate" action is required, it must run on every
  single PR, so I've removed the path filters.
- Remove the azure provider import from the provisioner facade, and
  build juju with BUILD_TAGS='minimal provider_lxd' in the migrate
  workflow. This shaves 3-4 minutes off the build time.
- Add a debug info step to the migrate workflow, so that if there are
  future failures, we can have a baseline to debug them.