feat: delete charm
The following deletes the charm from the database, along with all
the related data in the accompanying tables. Instead of creating
a function/method for each table, we can iterate through all the
knoew tables. If a table has no data it will be skipped over.

Note: we have to remove all the ancillary tables before the charm
table, which means removing the charm from the charm table is
done last.