feat: watch charm changes
This wires up the watch charm function, which includes wiring up
the generate triggers.

This also adds a test harness to the watcher tests to create a
watcher test that is predictable. The lifecycle of a watcher must
follow the following steps:

 1. Empty an initial event.
 2. The change stream should become idle.
 3. Run the setup for the test.
 4. The change stream should become idle.
 5. Assert the changes.

Steps from 3 to 5 are repeated for each test added to the harness.
Once the tests are run, the harness ensures that there are no more changes
before checking that the change stream is idle.
By ensuring that the change stream is idle, we can guarantee that the:

 1. The watcher has processed all the changes.
 2. There are no more changes to be processed.

This isolation technique allows us to test the watcher in a predictable
manner.

----

The controller config watching was also moved to the test harness to
prove that it also works on the controller database.