fix: change stream test composition
The TestOneChange assumed it would do one and only one loop of the
stream. Except that every so often it wouldn't be killed in time
and it would start to do another loop. There were no changes to
process, so would get stuck on waiting for a time out and because
we only mock the term timeout, not the back-off strategy. The solution
for just this test is to allow it to mock against any time, because
we don't care for this test. Alternatively, it would be better to
pass in two clocks, one for the back-off and another for the other
common use cases. This would allow for better control over testing.

In addition to the test changes, to better aid debugging, I've
added the long needed traces support.