docs(block): remove doc/block.txt
The block.txt was out of date reguarding what the code should look
like and what commands to use to enable and disable the blocks.
The doc was written like a specification and we have a new location
for that information.

I've dispersed the information into better locations, either in
doc.go files or on methods that actually deal with the blocking
actions. This is more likely going to be kept up to date from a
developers perspective.