feat: remove generated version file
If you can't archive a folder locally, we no longer need to have
the ability to generate a version file. We will still read the
version file in the zip file if it's present, but we don't
generate it. Along with that, the jujuignore file is also gone.
The same principal applies, that was only ever present because
of how we archived files. If we're not creating archives then we
no longer need it.

By simplifying the deployment process, we're also reducing the
complexity of a charm. We don't need a .jujuignore file, thus we
don't potentially need a version file. That information could be
stored in the manifest.yaml.

Note: everything that applies to charms, also applies to bundles
deploying local charm directories. Charms must be prepackaged;
ideally with charmcraft pack.

We're also dropping dependencies as we no longer globbing for files
when attempting to archive directories.