feat(machine): add all StatusInfo fields and status data tables
Using CASCADE on uuid foreign keys on status tables because we don't
want to have to cleanup the status and the status data whenever we
remove a machine. Whenever a uuid is deleted on the referenced table,
CASCADE allows automatic deletion of affected rows in the child table
to keep referential integrity.