chore: version to core package
The version package in the root of juju can be moved to the core
package. The core package is a public API that we support between
all the components.

Moving of the version/helper script was moved to the scripts package.
This scripts package needs to be cleaned at some point, so it's
correct and up to date.