feat: write charm initial info
When dual writing a charm, we also need to write additional
information. This includes, source, revision and the name. The charm
name is the name in the metadata. We don't allow naming of a charm
to be different that then name in the metadata. Revision is always
required, it will be either -1 (not set) or greater and equal to 0.

The validation in the service now expects the name to be set and
that we have the source of the charm. The source can either be
local or charmhub.