test: improve github pre merge checks
Add a single job which collates all changes files and
sets flags when then determine which tests should run.

Instead of marking each test as Required which would need
admin access to the repo, add a job which passes or fails
once all required tests are finished or skipped. This
allows a PR to be used to update required tests.