refactor(apiserver): drop some state calls
Drop some state calls in apiserver stateauthenticator.

There were 2 places where we unnecessarily call out to state in
stateauthenticator. In both these cases, we simply queried a model uuid

Get the model uuid in different ways.

1) In AuthenticateLoginRequest, the model uuid is included in the function
   params. Wire this up to checkCreds instead and use this.

   NOTE: This fixes a bug we had not noticed. We were using systemstate
   to get the model uuid, which returns the uuid of the controller model
   instead of the model being logged into

2) In newAuthContext, we queried state for the controller model uuid.
   This obviously never changes, so instead we load this uuid in
   httpserverargs worker, and pass it through into NewAuthenticator

This means now the only places we make calls to systemstate are for
macaroon related reasons. Once this is backed with DQLite, we will be
able to drop state from authContext entirely.