feat: coerce error for every transaction
The coerce error function isn't being correctly enforced. It's
sometimes applied and sometimes skipped. To improve consistency
the state base now ensures that the resulting error is correctly
coerced.

The fallout from this is that all sql errors _must_ be handled in
the transaction. This forces all code to correctly handled at the
exact query it failed at. You can no longer skip that! This should
lead to better error messages!

---

The coerce error function exists to prevent the leaking of the
sql errors outside of the domain. We don't want another situation
of the apiserver errors packaging being used in workers or state.
The coerce error will just prevent you from checking via errors.Is
and errors.As, so the only alternative is via a strings check. This
is frowned upon at every level.

---

The code changes are to move the sql checking into the correct
locations, inside of the transactions, next to queries.