feat: add user.Name type to replace string
Add a new type to represent a username. Before this PR, juju user names
were represnted either as a tag, or as a string. The tag represented
them as two distinct parts, the domain, and the username. For example,
alastair@somewhere. The part after the @ is the domain. When the access
service was initally implemented, only the part before the domain was
used. i.e. just the user name. However, it was quickly relised that the
domain would also be needed, or external and internal users would get
lumped together when they should not be.

This meant that to pass a username into domain you had to do
userTag.Id(), and NOT userTag.Name(). Using the Name would cause issues
with external users that were often missed in testing.

Later on, handling of external users was explicity added to the access domain.
This meant that the external users had to be distingused from the local
users. This meant reparsing the username string inside the domain to
determin if there was a domain included and if it was an external
domain.

Both of these together meant that usernames could be a serious cause of
bugs.

To fix this, this PR adds a user.Name type that represents a username
and has a .IsLocal method. It enforces type correctness around
usernames and removes the need for a lot of checks on user names. There
is a function to convert a tag into a user.Name type.