feat: db error metrics
We often see db errors when bootstrapping or creating a model, so
to better understand that, I've added a metric to witness it. The
metric is simple at them moment, it just highlights that there
was a retryable error. We might want to expose the underlying error
type (not message string) in the future, but we have to be
careful about the cardinality.

To prevent having to change the transaction runner for an optional
dependency I'm drawing inspiration from the tracing setup. In that
vein, I'm passing the metrics through the context. Alternatively, we
could create the transaction runner in the db accessor, but I'm not
sure it's worth the churn. If this becomes a common pattern, we can
change the db accessor to intern the transaction runner.