fix: auth for external users
Auth for external user was broken because if the user is not local, we
check their username against the logged in user name from state, and the
local username we were checking it against was empty. This was because
of a check against local users that would set their account detials to
empty if they had no macaroons. The macaroons are not stored in the
AccountDetails of the external user in this case, and so the username
it was compared against was empty.

This fix sets the account details to the username and macaroons
(which may be empty).