fix(auth): fix external user auth
When establishing api connections, we recently begun asserting that the
authenticated user name matches the name of the user in the store.

However, this revealed a bug with external users. In
NewAPIRootWithDialOpts, we would clear the account details
unnecessarily, including the username, if no macaroons are provided.

Change this to not clear the username.

This means we can collapse an if statement, as nowhere do we distinguish
between a 0 length array and nil for the Macaroons slice

A potential side effect of this would be to include the user Tag in
apiInfo in connectionInfo. However, we are saved here because we only
set the tag for local users.