chore: merge application and charm domain
The application domain needs to know about charms, so rather than
duplicating lots of complicated logic for encoding and decoding,
we should just merge them.

This was discussed with @manadart and although this could be
perceived as a god domain, although an application is the
manifestation of a charm, and a unit is a running instance of an
application. As all three need to know about each other, they're
very much intangled.