test: remove machine agent authorized keys test
Removes the cmd level package integration tests for authorized keys on
the agent. We have bash tests that cover this case now. This has been
written as an integration test in go at the cmd layer and as such has a
complicated lifecycle that is hard to break out of the apiserver suite.

We don't require this test in this form anymore.