fix(bootstrap): fix bug causing bootstrap to fail
We sometimes we've been seeing "bakery config already intialised"
errors when bootstraping from the bootstrap worker.

It seems like this is caused by errors further down the workers, causing
it to restart.

The bootstrap worker needs to be idempotent. If we see the bakery has
already been seeded, we should just continue.

Update doc.go to include the new aspects of state that are seeded, as
well as a note to ensure seeding processes are idempotent