test: unit test bad context id
In order to prove that we've got a bad context ID error, we have
to contrive a scenario, so that when the planets align the right
way, we get the correct error message.

This involves forcing the run action to wait when it's being
processed as a remote exection. Then we can call the Jujuc.Main
directly using the socket, which then allows us to execute a
bad request. A token generator was added to allow us to bypass
the token comparison for remote executions. Once all of this is
done, we then collapse the server and allow the action to complete.

---

I believe this is the first time someone has taken the time to
execute this code path outside of integration tests. We should
execute this path in unit tests more, either by making the runner
more composable, or implementing a better test harness.