feat: add external user on login if they do not exist
In mongo, external users did not exists in jujus database, only their
permissions did. With the switch to a relational schema, there are now
referential integrity checks against the users permissions, meaning the
user must exist in the database.

The issue comes when an user unknown to juju logs in. If the special
everyone@external user has permissions, this user should get them. These
permission checks have been added in previous PRs, but there are still
checks against the users name when they are searched in the database.
This leads to user not found errors.

With this change, if an unknown user logs in and needs to inherit the
external users permissions then a new user will now be created in the
database. This user will have no permissions of their own.

Revert "fix: allow unknown external users"

This reverts commit 28c780ea6f44c17a143e7ae95fb9b3f1feed447a.