chore(macaroon): bump macaroon-bakery patch + refactor
The latest patch version of go-macaroon-bakery/macaroon-bakery contains
a new constructor for RootKeyStores, which takes in a ContextBacking as
a param.

This means we no longer need to implement stub methods for Backing.

Make use of this new constructor, remove all stub methods + casts, and
make any refactors resulting from this