feat: wire through ssh importer client
This commit creates a new HTTP client for use in importing public ssh
keys and wires it through the manifolds and api server so the keymanager
facade can get access to it for importing public ssh keys on behalf of
users.

In this commit I have also fixed an issue with facade model context when
you request a http client. When requesting a http client you specify a
purpose for the http client and based on the purpose a pre injected http
client is returned. However if no client exists for the purpose the
context just returns nil. This is dangerous as it makes people assume
that because there is no error they get back a safe type.

Now HTTPClient on the context returns an error if it can not provide a
http client for the purpose so the caller knows not to go any further.