fix: idempotent binary storage
When adding the same binary hash twice, if the hash already exists,
we should just handle the error message and still write the binary
storage metadata.

Ideally, we would handle the error message outside of the state
layer, but this will be re-written at some point. For now, just allow
migrations to pass.