fix: launchpad username expectations
When re-implementing the user ssh key importer in native go for Juju we
put the expectation on the user that they would include the prefixed ~
character with their username when importing ssh keys from launch pad.

This was a breaking change for Juju version pre 4. This change now adds
it on behalf of the user and doesn't expect it in the subject name.