feat(access): add GetModelUsers to access service and pipe through
To get user info about a model, the apiserver would get all the users
from the model, then indavidually retrieve some extra info about each of
them. This happened in both the model manager and the user manager.

This change adds a method to the access service to replace this
functionality, getting all the required information from the database in
one hit.